backend/api/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
colorama==0.4.6
distlib==0.3.9
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
filelock==3.16.1
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.7
matplotlip==0.2
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.4
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
setuptools==75.6.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.28.0
